refactor(lib): add common utils
